Move key-related types to top-level module
