+++ identity_core/src/did/mod.rs
@@ -1,9 +1,13 @@
+mod authentication;
 #[allow(clippy::module_inception)]
 mod did;
 mod document;
 mod parser;
-mod types;
+mod proof;
+mod service;
 
+pub use authentication::*;
 pub use did::*;
 pub use document::*;
-pub use types::*;
+pub use proof::*;
+pub use service::*;
