+++ identity_core/src/key/mod.rs
@@ -1,15 +1,11 @@
-mod authentication;
 mod key_data;
 mod key_index;
 mod key_relation;
 mod key_type;
 mod public_key;
-mod service;
 
-pub use authentication::*;
 pub use key_data::*;
 pub use key_index::*;
 pub use key_relation::*;
 pub use key_type::*;
 pub use public_key::*;
-pub use service::*;
